+++ tests_corpus/valid/primaruy-key-collation-sorting.sql
@@ -1,0 +1,5 @@
+CREATE TABLE person(
+    fullname NOT NULL,
+    birthday DEFAULT NULL,
+    PRIMARY KEY (fullname COLLATE BINARY ASC)
+);