+++ tests_corpus/valid/primaruy-key-inline-collation-sorting.sql
@@ -1,0 +1,4 @@
+CREATE TABLE person(
+    fullname NOT NULL PRIMARY KEY ASC COLLATE BINARY,
+    birthday DEFAULT NULL
+);